rename: event -> Notifier 로 이름 수정
