+++ layer-external/src/main/java/org/layer/discord/notifier/CreateRetrospectEventNotifier.java
@@ -1,6 +1,7 @@
-package org.layer.discord.event;
+package org.layer.discord.notifier;
 
 import org.layer.discord.DiscordAppender;
+import org.layer.event.retrospect.CreateRetrospectEvent;
 import org.springframework.context.annotation.Profile;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
@@ -10,13 +11,15 @@ import lombok.RequiredArgsConstructor;
 @Component
 @Profile("prod")
 @RequiredArgsConstructor
-public class ErrorEventNotifier {
+public class CreateRetrospectEventNotifier {
+
 	private final DiscordAppender discordAppender;
 
 	@EventListener
-	public void handleSignUpEvent(ErrorEvent event) {
-		discordAppender.createErrorAppend(
-			event.message(),
-			event.stackTrace());
+	public void handleSignUpEvent(CreateRetrospectEvent event) {
+		discordAppender.createRetrospectAppend(
+			event.title(),
+			event.memberId(),
+			event.createdDate());
 	}
 }
