+++ layer-external/src/main/java/org/layer/discord/notifier/CreateSpaceEventNotifier.java
@@ -1,7 +1,7 @@
-package org.layer.discord.event;
+package org.layer.discord.notifier;
 
 import org.layer.discord.DiscordAppender;
-import org.layer.event.member.SignUpEvent;
+import org.layer.event.space.CreateSpaceEvent;
 import org.springframework.context.annotation.Profile;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
@@ -11,13 +11,14 @@ import lombok.RequiredArgsConstructor;
 @Component
 @Profile("prod")
 @RequiredArgsConstructor
-public class SignUpEventNotifier {
+public class CreateSpaceEventNotifier {
+
 	private final DiscordAppender discordAppender;
 
 	@EventListener
-	public void handleSignUpEvent(SignUpEvent event) {
-		discordAppender.createMember(
-			event.name(),
+	public void handleSignUpEvent(CreateSpaceEvent event) {
+		discordAppender.createSpaceAppend(
+			event.title(),
 			event.memberId(),
 			event.eventTime());
 	}
