+++ layer-external/src/main/java/org/layer/discord/notifier/ErrorEventNotifier.java
@@ -1,7 +1,6 @@
-package org.layer.discord.event;
+package org.layer.discord.notifier;
 
 import org.layer.discord.DiscordAppender;
-import org.layer.event.space.CreateSpaceEvent;
 import org.springframework.context.annotation.Profile;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
@@ -11,15 +10,13 @@ import lombok.RequiredArgsConstructor;
 @Component
 @Profile("prod")
 @RequiredArgsConstructor
-public class CreateSpaceEventNotifier {
-
+public class ErrorEventNotifier {
 	private final DiscordAppender discordAppender;
 
 	@EventListener
-	public void handleSignUpEvent(CreateSpaceEvent event) {
-		discordAppender.createSpaceAppend(
-			event.title(),
-			event.memberId(),
-			event.eventTime());
+	public void handleSignUpEvent(ErrorEvent event) {
+		discordAppender.createErrorAppend(
+			event.message(),
+			event.stackTrace());
 	}
 }
