+++ layer-external/src/main/java/org/layer/discord/notifier/SignUpEventNotifier.java
@@ -1,6 +1,7 @@
-package org.layer.discord.event;
+package org.layer.discord.notifier;
 
 import org.layer.discord.DiscordAppender;
+import org.layer.event.member.SignUpEvent;
 import org.springframework.context.annotation.Profile;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
@@ -10,11 +11,14 @@ import lombok.RequiredArgsConstructor;
 @Component
 @Profile("prod")
 @RequiredArgsConstructor
-public class MemberActivityEventNotifier {
+public class SignUpEventNotifier {
 	private final DiscordAppender discordAppender;
 
 	@EventListener
-	public void handleSignUpEvent(MemberActivityEvent event) {
-		discordAppender.aggregateMemberActivity(event.activities());
+	public void handleSignUpEvent(SignUpEvent event) {
+		discordAppender.createMember(
+			event.name(),
+			event.memberId(),
+			event.eventTime());
 	}
 }
